4.23 finished, need redo

+++ part4/blogList/controllers/blogs.js
@@ -5,28 +5,24 @@ const jwt = require('jsonwebtoken')
 
 notesRouter.get('/', async (request, response) => {
   const blogs = await Blog
-    .find({}).populate('user', { username: 1, name:1})
+    .find({}).populate('user', { username: 1, name: 1 })
 
   response.json(blogs)
 })
 
 notesRouter.post('/', async (request, response) => {
   const body = request.body
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  const user = request.user
 
-  if (!request.token || !decodedToken.id) {
-    return response.status(401).json({ error: 'token missing or invalid' })
-  }
-  const user = await User.findById(decodedToken.id)
-
+ 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
-    likes:body.likes,
-    user: user._id  
+    likes: body.likes,
+    user: user._id
   })
-  
+
   const savedBlog = await blog.save()
   user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
@@ -35,8 +31,14 @@ notesRouter.post('/', async (request, response) => {
 })
 
 notesRouter.delete('/:id', async (request, response, next) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  
+  const user = request.user
+  const blogsFromUser = user.blogs
+  
+  if (blogsFromUser.includes(request.params.id)){
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
 })
 
 notesRouter.put('/:id', async (request, response, next) => {
